+++ osidb/signals.py
@@ -8,7 +8,6 @@ from django.template.loader import render_to_string
 from django.utils import timezone
 from jira import JIRA
 
-from apps.workflows.workflow import WorkflowModel
 from collectors.jiraffe.constants import HTTPS_PROXY
 from config.settings import EmailSettings
 from osidb.helpers import get_env, get_execution_env
@@ -159,10 +158,7 @@ def updated_local_updated_dt_affectcvss(sender, instance, **kwargs):
 
 @receiver(post_save, sender=Affect)
 def create_flaw_labels(sender, instance, **kwargs):
-    if (
-        instance.flaw.workflow_state
-        == WorkflowModel.WorkflowState.PRE_SECONDARY_ASSESSMENT
-    ):
+    if instance.flaw.workflow_state == "PRE_SECONDARY_ASSESSMENT":
         if instance._state.adding:
             FlawCollaborator.objects.create_from_affect(instance)
         else:
@@ -171,10 +167,7 @@ def create_flaw_labels(sender, instance, **kwargs):
 
 @receiver(post_delete, sender=Affect)
 def delete_flaw_labels(sender, instance, **kwargs):
-    if (
-        instance.flaw.workflow_state
-        == WorkflowModel.WorkflowState.PRE_SECONDARY_ASSESSMENT
-    ):
+    if instance.flaw.workflow_state == "PRE_SECONDARY_ASSESSMENT":
         FlawCollaborator.objects.mark_irrelevant(instance.flaw)
 
 
@@ -182,10 +175,9 @@ def delete_flaw_labels(sender, instance, **kwargs):
 def create_labels_on_promote(sender, instance, **kwargs):
     if (
         not instance._state.adding
-        and instance.workflow_state
-        == WorkflowModel.WorkflowState.PRE_SECONDARY_ASSESSMENT
+        and instance.workflow_state == "PRE_SECONDARY_ASSESSMENT"
         and Flaw.objects.get(pk=instance.pk).workflow_state
-        != WorkflowModel.WorkflowState.PRE_SECONDARY_ASSESSMENT
+        != "PRE_SECONDARY_ASSESSMENT"
     ):
         FlawCollaborator.objects.create_from_flaw(instance)